Refactored transaction limit tests
